+++ src/index.js
@@ -1,10 +1,9 @@
 
 import "./style.css";
 
-// import "./js/classCard";
-// import "./js/api";
-// import "./js/classCardList";
-// import "./js/classPopup";
-// import "./js/validationForms";
-import "./js/script-seven-sprint";
-import "./js/script-eight-sprint";
+import "./js/classCard";
+import "./js/api";
+import "./js/classCardList";
+import "./js/classPopup";
+import "./js/validationForms";
+
